Fix Day 13

+++ src/day13.rs
@@ -1,15 +1,15 @@
 use aoc_runner_derive::{aoc, aoc_generator};
 
-const MAX_TIMES_PRESSED: usize = 100;
-const PUSH_A_COST: usize = 3;
-const PUSH_B_COST: usize = 1;
-const EXTRA_DISTANCE: usize = 10_000_000_000_000;
+const MAX_TIMES_PRESSED: i64 = 100;
+const PUSH_A_COST: i64 = 3;
+const PUSH_B_COST: i64 = 1;
+const EXTRA_DISTANCE: i64 = 10_000_000_000_000;
 
 #[derive(Debug)]
 struct MachineSetup {
-    button_a: (usize, usize),
-    button_b: (usize, usize),
-    prize_location: (usize, usize),
+    button_a: (i64, i64),
+    button_b: (i64, i64),
+    prize_location: (i64, i64),
 }
 
 #[aoc_generator(day13)]
@@ -17,9 +17,9 @@ fn parse_input(input: &str) -> Vec<MachineSetup> {
     use aoc_parse::{parser, prelude::*};
 
     let parser = parser!(sections(
-    button_a:line("Button A: X+" usize ", Y+" usize)
-    button_b:line("Button B: X+" b_x:usize ", Y+" b_y:usize)
-    prize_location:line("Prize: X=" prize_x:usize ", Y=" prize_y:usize)
+    button_a:line("Button A: X+" i64 ", Y+" i64)
+    button_b:line("Button B: X+" b_x:i64 ", Y+" b_y:i64)
+    prize_location:line("Prize: X=" prize_x:i64 ", Y=" prize_y:i64)
      => MachineSetup{
          button_a,
          button_b,
@@ -36,7 +36,7 @@ fn solve_machine(
         prize_location,
     }: &MachineSetup,
     with_limit: bool,
-) -> Option<usize> {
+) -> Option<i64> {
     let a_nom = prize_location.0 * button_b.1 - prize_location.1 * button_b.0;
     let b_nom = button_a.0 * prize_location.1 - button_a.1 * prize_location.0;
     let denom = button_a.0 * button_b.1 - button_a.1 * button_b.0;
@@ -56,7 +56,7 @@ fn solve_machine(
 }
 
 #[aoc(day13, part1)]
-fn part1(machines: &[MachineSetup]) -> usize {
+fn part1(machines: &[MachineSetup]) -> i64 {
     machines
         .iter()
         .filter_map(|machine_setup| solve_machine(machine_setup, true))
@@ -64,7 +64,7 @@ fn part1(machines: &[MachineSetup]) -> usize {
 }
 
 #[aoc(day13, part2)]
-fn part2(machines: &[MachineSetup]) -> usize {
+fn part2(machines: &[MachineSetup]) -> i64 {
     machines
         .iter()
         .filter_map(
